--- src/services/invoice-invoiceDetail.interface.ts
@@ -1,10 +1,0 @@
-export interface IInvoice {
-  invoiceId: number;
-  date: Date;
-  customerName: string;
-  invoiceDetailId: number;
-  name: string;
-  description: string;
-  quantity: number;
-  price: number;
-}
